+++ CHANGELOG.md
@@ -3,7 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
-## [v1.0-1.16.5] - 2021-07-03
+## [v1.0.1-1.16.5] - 2021-10-24
+## Changed
+- Search box can be cleared via right-click
+- Search box is more apparent when empty
+
+## [v1.0-1.16.5] - 2021-10-21
 - Initial release
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/